requirements.txt
pyrogram>=2.0.0
TgCrypto>=1.2.5
python-dotenv>=1.0.0
pymongo>=4.6.0
py7zr>=0.20.0
rarfile>=4.1
qrcode[pil]>=8.0
pillow>=10.0.0
aiofiles>=23.0.0
